requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
Django==6.0.1
django-cors-headers==4.9.0
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
et_xmlfile==2.0.0
gunicorn==24.1.1
numpy==2.4.1
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
qrcode==8.2
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0
xlrd==2.0.2
